Always ask if we can join if we are not reloading the page
Note: This uses the browsers sessionStorage not the browserStorage.
As per https://stackoverflow.com/q/20325763 this is NOT shared between tabs.
For us this is the solution we were looking for, as it allows us to
identify if a user reloaded a conversation in the same tab,
or entered it in another tab.

Signed-off-by: Joas Schilling <coding@schilljs.com>