Offload generation of blurred backgrounds to a worker
Until now the generation of the blurred backgrounds was done in a normal
canvas, so it happened in the main thread and made the UI unresponsive
while the image was being generated. Now, to solve that, the blurred
background is generated in an OffscreenCanvas in a worker (if
supported by the browser).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>